requirements.txt
fastapi[all]
uvicorn
pandas
scikit-learn==1.4.1.post1
joblib
pydantic
catboost
pytest
pyyaml
logging.config
